cms_bot/pyproject.toml
[tool.poetry]
name = "dcms-bot"
version = "0.1.0"
description = ""
authors = ["Charles de Freitas <me@chasbob.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
sqlalchemy = "^1.3.19"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
